Added OAuth to the extension
The code in background gets the authentication token which is sent with the youtube search request through a modification of inject.js. Chrome messaging is used to connect the content to the background. In addition the extension was given a key to tie it to an Google API Application. This should change the id of the extension the next time it is loaded.

I am not certain that the OAuth user is going to solve the quote issue. The API key has been changed to tie it to the same account as the OAuth account. We will have to see. I can request a higher quota if required. Otherwise research on this will need to be done. However the OAuth code is useful for future features.