Add extension icon
Added a new icon in multiple sizes that reflects the products indended purpose to improve meet sessions in school. The icon enhances the Meet icon with a mortarboard.

Also removed the browser action section from the manifest and removed the codday icon. It would be more appropriate to use a page action than a browser action if we need to have an icon on the browser.

Updated .gitignore to ignore any .DS_Store files